Improved twitter messages
Now is easier than before to get the information and store it in the same file json.
It is pending implement obtaining language in the pipe.
It also gives a problem recognizing some tweets, a tryCatch has been placed to be able to follow the execution if that error occurs.