Created the "process_files" function and improvements in the pipes
A function has been created that allows to pass a path of the files and a class that inherits TypePipe which contains the pipes that have to be executed in the pipeAll method.
In addition, it has improved the pipes which accessed the resource files and the pipe that formed the data frame.