Created the structure to control the dependencies between pipes
Added control of pipes that needed to be run before a concrete pipe and pipes can not be executed after the concrete pipe has been executed.
This structure has been introduced in all pipes.
It is necessary to specify the dependencies of each pipe, which can be indicated by default in the constructor or the user can enter the name of the pipes in the dependency lists when he makes a new pipe.