First version of a class and methods for user
The Bdp4R class has been created that will allow the user to process the
files. In addition, a function has been created that allows the user not
not to create the object.

The variables used with <<- have been changed to attributes of the
newly created class and will act as static attributes of the class,
allowing easy access throughout to the preprocessing.